+++ src/routes/(root)/opi/ko/quick-guide/payment/index.mdx
@@ -7,8 +7,6 @@ versionVariants:
   v1: /opi/ko/integration/ready/readme?v=v1
 ---
 
-import Hint from "~/components/Hint";
-
 import { Condition, InteractiveDoc, Section } from "./_preview";
 
 {/* client:import-portone-sdk
@@ -25,6 +23,14 @@ client:handle-payment-status:paid */}
     ## 포트원 브라우저 SDK 불러오기
 
     포트원 브라우저 SDK를 불러옵니다.
+
+    <Condition language="frontend/react">
+      아래 명령어로 브라우저 SDK를 설치합니다.
+
+      ```shell
+      npm install @portone/browser-sdk
+      ```
+    </Condition>
   </Section>
 
   <Section section="client:fetch-item">
@@ -43,10 +49,6 @@ client:handle-payment-status:paid */}
     ## Payment ID 설정
 
     Payment ID는 결제 요청 시 사용하는 고유한 ID입니다. 포트원에서 제공하지 않고 수동으로 생성해야 합니다.
-
-    <Hint style='warning'>
-      결제대행사에 따라 Payment ID의 형식에 제한이 있을 수 있습니다.
-    </Hint>
   </Section>
 
   <Condition when={({ pg: { name } }) => ['ksnet', 'inicis', 'smartro'].includes(name)}>
@@ -99,6 +101,14 @@ client:handle-payment-status:paid */}
     ## 포트원 서버 SDK 불러오기
 
     포트원 서버 SDK를 불러옵니다.
+
+    <Condition language="backend/express">
+      아래 명령어로 서버 SDK를 설치합니다.
+
+      ```shell
+      npm install @portone/server-sdk
+      ```
+    </Condition>
   </Section>
 
   <Section section="server:complete-payment">
